requirements.txt
aiohttp==3.5.4
psycopg2-binary==2.8.3
aiopg==1.0.0
PyJWT==1.7.1
peewee==3.10.0
peewee-async==0.6.0a
peewee-migrations==0.3.18
